Fixes for bug dot(x,y). Which led to a rabbithole of changes. All tests pass.
added Node <: Number

This caused make_Expr and function_body! to fail. changed Dict to IdDict in code generation. This caused fields in DerivativeGraph to be redefined to use IdDict.

sparse calls iszero on a node object which cause a type promotion error. Added Base.iszero function(a::Node)

got type ambiguity on ^ so added specialized version for (Node,Integer)